Update recommended installation procedure.
This commit adss a requirements.txt file and updates the recommended
routing as described in the README file using pip.

The setuptools.setup() function does not correctly parse the package and
version requirements and is unable to detect universal wheels
(compatible with Python 2 and 3) meeting any depedency requirements.
Using pip all latest versions and depedencies are correctly installed.